% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotLinear.R
\name{plotLinear}
\alias{plotLinear}
\title{Plot (improper) weighted linear combinations of variables}
\usage{
plotLinear(formula, data, scaling = "none", weighting = 1)
}
\arguments{
\item{formula}{an object of class \code{"formula"} (or one that can
be coerced to that class): a symbolic
description of the model to be fitted.}

\item{data}{an optional data frame, list or environment
(or object coercible by \code{as.data.frame}
to a data frame) containing the variables in the model.}

\item{scaling}{method of scaling to be used. Possible choices
are: "none", "zscore", "minmax", "medmad" (see
\code{\link{ilm}} for learning more)}

\item{weighting}{weights to be applied to all the \emph{variables}
                 (notice: this is different than \code{weights}
                 parameter in \code{\link{lm}}).}
}
\description{
Plot (improper) weighted linear combinations of variables
}
\seealso{
\code{\link{ilm}}
}

