% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/S3methods.R
\name{weighting}
\alias{weighting}
\title{Extract variable weights from ilm or iglm object}
\usage{
weighting(object)
}
\arguments{
\item{object}{an \code{ilm} or \code{iglm} object.}
}
\description{
Extract variable weights from ilm or iglm object
}
\seealso{
\code{\link{ilm}}, \code{\link{iglm}}
}

